В процессе создания валидации
